UPBGE: Use int instead of short in vertex array index.
It fixes some issues about half converted meshes cause of an huge number of vertexes.

+++ source/gameengine/Rasterizer/RAS_MaterialBucket.h
@@ -85,7 +85,7 @@ class RAS_DisplayArray
 {
 public:
 	vector<RAS_TexVert> m_vertex;
-	vector<unsigned short> m_index;
+	vector<unsigned int> m_index;
 
 	enum {BUCKET_MAX_INDEX = 65535};
 	enum {BUCKET_MAX_VERTEX = 65535};
